+++ tests/fast/utils/test_arguments.py
@@ -9,11 +9,8 @@ from unittest.mock import patch
 
 import pytest
 
-from miles.backends.sglang_utils.arguments import validate_args as sglang_validate_args
-from miles.backends.sglang_utils.sglang_engine import _compute_server_args
 from miles.utils.arguments import (
     _maybe_apply_dumper_overrides,
-    _maybe_enable_true_on_policy_sglang_cp_lm_head,
     get_miles_extra_args_provider,
 )
 from miles.utils.misc import function_registry
@@ -153,130 +150,3 @@ def test_recompute_logprobs_via_prefill_flag_is_parsed():
     assert args.recompute_logprobs_via_prefill is True
 
 
-@pytest.mark.parametrize(
-    ("true_on_policy_mode", "sglang_attn_cp_size", "sglang_attention_context_parallel_size", "expected_dp_lm_head"),
-    [
-        (False, 4, 4, False),
-        (True, 1, 1, False),
-        (True, 4, 1, False),
-        (True, 1, 4, False),
-    ],
-)
-def test_true_on_policy_sglang_cp_preserves_dp_lm_head_default(
-    true_on_policy_mode: bool,
-    sglang_attn_cp_size: int,
-    sglang_attention_context_parallel_size: int,
-    expected_dp_lm_head: bool,
-):
-    args = SimpleNamespace(
-        true_on_policy_mode=true_on_policy_mode,
-        sglang_attn_cp_size=sglang_attn_cp_size,
-        sglang_attention_context_parallel_size=sglang_attention_context_parallel_size,
-        sglang_enable_dp_lm_head=False,
-    )
-
-    _maybe_enable_true_on_policy_sglang_cp_lm_head(args)
-
-    assert args.sglang_enable_dp_lm_head is expected_dp_lm_head
-
-
-@pytest.mark.parametrize(
-    (
-        "rollout_num_gpus_per_engine",
-        "sglang_attention_context_parallel_size",
-        "recompute_logprobs_via_prefill",
-        "expected_target",
-        "expected_prefill_only",
-        "expected_dp_lm_head",
-    ),
-    [
-        (1, 1, False, "fsdp", False, False),
-        (4, 1, True, "fsdp_tp", True, False),
-        (8, 4, True, "fsdp_tp", True, False),
-    ],
-)
-def test_true_on_policy_args_propagate_to_sglang_server_args(
-    rollout_num_gpus_per_engine: int,
-    sglang_attention_context_parallel_size: int,
-    recompute_logprobs_via_prefill: bool,
-    expected_target: str,
-    expected_prefill_only: bool,
-    expected_dp_lm_head: bool,
-):
-    args = SimpleNamespace(
-        rollout_num_gpus_per_engine=rollout_num_gpus_per_engine,
-        sglang_data_parallel_size=1,
-        sglang_pipeline_parallel_size=1,
-        sglang_expert_parallel_size=1,
-        sglang_attention_context_parallel_size=sglang_attention_context_parallel_size,
-        sglang_enable_dp_attention=False,
-        sglang_enable_dp_lm_head=False,
-        sglang_router_policy=None,
-        sglang_router_ip=None,
-        true_on_policy_mode=True,
-        recompute_logprobs_via_prefill=recompute_logprobs_via_prefill,
-        sglang_rl_on_policy_target=None,
-        sglang_true_on_policy_contract="qwen3_dense_true_on_policy_v1",
-        sglang_enable_deterministic_inference=False,
-        sglang_enable_prefill_only_deterministic_inference=False,
-        hf_checkpoint="hf://dummy",
-        seed=7,
-        offload_rollout=False,
-        num_gpus_per_node=8,
-        use_rollout_routing_replay=False,
-        fp16=False,
-    )
-
-    sglang_validate_args(args)
-
-    server_args, _ = _compute_server_args(
-        args,
-        rank=0,
-        dist_init_addr="127.0.0.1:12345",
-        nccl_port=12346,
-        host="127.0.0.1",
-        port=30000,
-    )
-
-    assert args.sglang_rl_on_policy_target is None
-    assert args.sglang_enable_deterministic_inference is True
-    assert args.sglang_enable_prefill_only_deterministic_inference is expected_prefill_only
-    assert args.sglang_enable_dp_lm_head is expected_dp_lm_head
-    assert server_args["rl_on_policy_target"] is None
-    assert server_args["true_on_policy_contract"] == "qwen3_dense_true_on_policy_v1"
-    assert server_args["enable_deterministic_inference"] is True
-    assert server_args["enable_prefill_only_deterministic_inference"] is expected_prefill_only
-    assert server_args["enable_dp_lm_head"] is expected_dp_lm_head
-
-
-def test_true_on_policy_sglang_cp_dp_lm_head_overrides_engine_defaults():
-    args = SimpleNamespace(
-        rollout_num_gpus_per_engine=8,
-        sglang_data_parallel_size=1,
-        sglang_pipeline_parallel_size=1,
-        sglang_expert_parallel_size=1,
-        sglang_dp_size=1,
-        sglang_pp_size=1,
-        sglang_ep_size=1,
-        sglang_attention_context_parallel_size=4,
-        sglang_enable_dp_lm_head=True,
-        true_on_policy_mode=True,
-        hf_checkpoint="hf://dummy",
-        seed=7,
-        offload_rollout=False,
-        num_gpus_per_node=8,
-        use_rollout_routing_replay=False,
-        fp16=False,
-    )
-
-    server_args, _ = _compute_server_args(
-        args,
-        rank=0,
-        dist_init_addr="127.0.0.1:12345",
-        nccl_port=12346,
-        host="127.0.0.1",
-        port=30000,
-        sglang_overrides={"enable_dp_lm_head": False},
-    )
-
-    assert server_args["enable_dp_lm_head"] is True
